+++ test/Tests/IntMap.hs
@@ -5,6 +5,7 @@
 module Tests.IntMap (intMapTests) where
 
 import           Control.Applicative
+import           Control.Comonad
 import           Data.Coerce
 import           Data.Foldable
 import           Data.Functor.Identity
@@ -109,6 +110,26 @@ prop_splitRoot = property $ do
       Nothing          -> True
       Just ys@(y :| _) -> x < y && ascending ys
 
+prop_extract_duplicate :: Property
+prop_extract_duplicate = property $ do
+    n <- forAll neIntMapGen
+    tripping n duplicate
+               (Identity . extract)
+
+prop_fmap_extract_duplicate :: Property
+prop_fmap_extract_duplicate = property $ do
+    n <- forAll neIntMapGen
+    tripping n duplicate
+               (Identity . fmap extract)
+
+prop_duplicate_duplicate :: Property
+prop_duplicate_duplicate = property $ do
+    n <- forAll neIntMapGen
+    let dd1 = duplicate . duplicate      $ n
+        dd2 = fmap duplicate . duplicate $ n
+    assert $ NEM.valid dd1
+    assert $ NEM.valid dd2
+    dd1 === dd2