+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Version 0.1.1.0
+---------------
+
+<https://github.com/mstksg/nonempty-containres/releases/tag/v0.1.1.0>
+
+*   `Comonad` instances added for `Map k` and `IntMap`, based on [Faucelme's
+    suggestion][comonad]
+
+[comonad]: https://www.reddit.com/r/haskell/comments/a1qjcy/nonemptycontainers_nonempty_variants_of/eat5r4h/
+
 Version 0.1.0.0
 ---------------